Another implementation of trees: as sorted alists.
HOL4 has a few implementations of partial maps as lookup trees
already. This one is faster than balanced_btree, because the
sorting and balancing decisions are largely made outside the
logic. Inserting a key/value pair, for instance, requires log-n
insertion and balancing steps, and 1-2 comparisons to show that
the inserted key is correctly placed compared to its neighbours.

The 'external' interface allows other partial functions to be
registered as having an assoc-list representation, and exports
a conversion which evaluates those functions.