Committing before major change.
I've run through backendProofScript a lot using the main oracle as
constructed from combinators that match the way it's later destructed
in the semantics proofs. The problem is, this beast is hard to describe.
I've set things up to move to a definite oracle from which the oracles
of the various sublanguages are projections. Committing before destroying
prior work.